Move normalizeWheel into view-base
